requirements.txt
about-time==4.2.1
alive-progress==3.1.5
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.6.0
arrow==1.3.0
attrs==23.1.0
binaryornot==0.4.4
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==2.1.1
cli-artifacts==1.6.4
cli-code-quality==1.0.4
cli-dla==1.2.2
cli-load-shedding==0.2.1
cli-performance==1.10.3
cli-ratelimit==0.8.0
cli-sandbox==1.2.10
cli-scm==1.0.12
cli-services==1.19.11
click==8.1.7
click-plugins==1.1.1
click-spinner==0.1.10
code-quality-parser==1.0.2
cookiecutter==2.1.1
cryptography==41.0.6
dacite==1.8.1
decorator==5.1.1
distlib==0.3.7
docker==6.1.3
exceptiongroup==1.2.0
filelock==3.13.1
furycli==3.2.6
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
gitdb==4.0.11
GitPython==3.1.41
grapheme==0.6.0
grpcio==1.60.1
grpcio-tools==1.60.1
identify==2.5.31
idna==3.4
importlib-resources==6.1.1
iniconfig==2.0.0
Jinja2==3.1.2
jinja2-time==0.2.0
jsonpickle==3.0.2
jsonschema==4.19.1
jsonschema-specifications==2023.11.1
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
nodeenv==1.8.0
numpy==1.26.3
oauthlib==3.2.2
package-name==0.1
packaging==23.2
pandas==2.2.0
pinject==0.14.1
platformdirs==3.11.0
pluggy==1.3.0
portpicker==1.6.0
pre-commit==3.5.0
protobuf==4.25.3
psutil==5.9.6
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
PyJWT==2.8.0
pyperclip==1.8.2
pytest==7.4.4
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.28.1
rich==13.7.0
rpds-py==0.12.0
semver==3.0.2
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.1
tabulate==0.9.0
text-unidecode==1.3
tiger-python-helper==0.8.3
tomli==2.0.1
tqdm==4.66.1
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.0
virtualenv==20.24.6
websocket-client==1.6.4
zipp==3.17.0
